requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
anthropic>=0.25.0
pdfplumber>=0.11.0
textstat>=0.7.3
pydantic>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
